move FileUpload one level up
